Automatically remove unwatched fds from the internal state
This frees up memory.

It also fixes a bug: if close() is called on the fd, the internal state
does not reflect this, but epoll itself automatically unregisters it.
This is problematic because if the fd is reused, then it won't be
re-registered because our internal state says that it already has been.